First draft: all features check, commit for deployment

+++ routes/employeeRoutes.js
@@ -1,6 +1,11 @@
+const { validationResult, matchedData, checkSchema, oneOf, body } = require('express-validator');
+const createEmployeeValidationSchema = require("../utils/createEmployeeValidationSchema.js");
+const updateEmployeeValidationSchema = require("../utils/updateEmployeeValidationSchema.js");
+
 const express = require('express');
 const router = express.Router();
 const Employee = require('../model/employee.js');
+const { isValidObjectId } = require('mongoose');
 
 // route: GET /api/v1/emp/employees
 // get all employees
@@ -23,9 +28,34 @@ router.get('/employees', async (req, res) => {
 
 // route: POST /api/v1/emp/employees
 // create employee
-router.post('/employees', async (req, res) => {
+router.post('/employees', 
+    checkSchema(createEmployeeValidationSchema),
+    async (req, res) => {
     try {
-        const newEmployee = new Employee(req.body);
+        // check validation
+        const expressValidationResult = validationResult(req);
+        // if any error, return 400 and error msg
+        if(!expressValidationResult.isEmpty()){
+            return res.status(400).send({
+                message: "Oops, you've entered some invalid fields",
+                error: expressValidationResult.array()});
+        }
+
+        // use validated data to create employee
+        const data = matchedData(req);
+        const { first_name, last_name, email, position, 
+            salary, date_of_joining, department } = data;
+        
+        const newEmployee = new Employee({
+            first_name: first_name,
+            last_name: last_name,
+            email: email,
+            position: position,
+            salary: salary,
+            date_of_joining: date_of_joining,
+            department: department
+        });
+
         await newEmployee.save(); // persist to db
 
         // on success
@@ -50,15 +80,19 @@ router.post('/employees', async (req, res) => {
 // get specific employee by id
 router.get('/employees/:empID', async (req, res) => {
     try {
+        // check id format matches mongodb _id
         const empID = req.params.empID;
+        if(!isValidObjectId(empID))
+            return res.status(400).send({status: false, message: `Employee ID is invalid.`});
+        
+        // query for employee
         const employee = await Employee.findOne({_id: empID});
         if(employee)
-            res.status(200).send(employee);
+            return res.status(200).send(employee);
         else
-            res.status(404).send({status: false, message: `Employee with id '${empID} cannot be found'`});
+            return res.status(404).send({status: false, message: `Employee with given id cannot be found.`});
 
     } catch (err) {
-
         res.status(500).send({ 
             message: "Employee retrieval failed", 
             status: 'Status 500: internal server error', 
@@ -70,15 +104,41 @@ router.get('/employees/:empID', async (req, res) => {
 
 // route: PUT /api/v1/emp/employees/eid
 // delete employee by id
-router.put('/employees/:eid', async (req, res) => {
+router.put('/employees/:eid', 
+    checkSchema(updateEmployeeValidationSchema),
+    async (req, res) => {
     try {
+        // check if id matches mongodb _id format
+        if(!isValidObjectId(req.params.eid))
+            return res.status(400).send({status: false, message: `Employee ID is invalid.`});
+        
+        // check express-validator results, exit if any errors
+        const expressValidationResult = validationResult(req);
+        if(!expressValidationResult.isEmpty()){
+            return res.status(400).send({
+                message: "Oops, you've entered some invalid fields",
+                error: expressValidationResult.array()});
+        }
 
+        // get validated data
+        const data = matchedData(req);
+        const { first_name, last_name, email, position, 
+            salary, date_of_joining, department } = data;
+        
         const employee = await Employee.findOneAndUpdate(
-            {_id: req.params.eid}, // identifier 
-            // $set determines which fields to update
+            {_id: req.params.eid},  
+            // $set indicates which fields to update
             { $set:
                 {
-                    ...req.body,
+                    // we could use '...req.body' here
+                    // but created_at and updated_at shouldn't be updateable
+                    first_name: first_name,
+                    last_name: last_name,
+                    email: email,
+                    position: position,
+                    salary: salary,
+                    date_of_joining: date_of_joining,
+                    department: department,
                     updated_at: Date.now() // update this field every time
                 }
             },
@@ -86,12 +146,12 @@ router.put('/employees/:eid', async (req, res) => {
             {new: true}
 
         );
-        const id = employee._id;
 
+        const id = employee._id;
         if(employee)
-            res.status(200).send({status: true, message: `Employee '${id}' updated successfully `});
+            return res.status(200).send({status: true, message: `Employee '${id}' updated successfully `});
         else
-            res.status(404).send({status: false, message: `Employee with id '${id}' cannot be found'`});
+            return res.status(404).send({status: false, message: `Employee with id '${id}' cannot be found'`});
 
     } catch (err) {
 
@@ -108,15 +168,21 @@ router.put('/employees/:eid', async (req, res) => {
 // delete employee by id
 router.delete('/employees', async (req, res) => {
     try {
-        const employee = await Employee.findOneAndDelete({_id: req.query.eid});
-        const id = employee._id;
-        if(employee)
-            res.status(204).send({status: true, message: `Employee '${id}' deleted successfully `});
-        else
-            res.status(404).send({status: false, message: `Employee with id '${id}' cannot be found'`});
+        // check format of id matches mongodb _id
+        if(!isValidObjectId(req.query.eid))
+            return res.status(400).send({status: false, message: `Employee ID is invalid.`});
+
+        // avoid error from findOneAndDelete() by first checking if employee exists 
+        const employee = await Employee.findOne({_id: req.query.eid});
+        if(!employee){
+            return res.status(404).send({status: false, message: `Employee with given id cannot be found.`});
+        } else {
+            // delete and return message
+            const deletedEmployee = await Employee.findOneAndDelete({_id: req.query.eid}); // delete employee
+            return res.status(204).send({status: true, message: `Employee '${deletedEmployee._id}' deleted successfully `});
+        }
 
     } catch (err) {
-
         res.status(500).send({ 
             message: "Employee delete failed", 
             status: 'Status 500: internal server error',
